+++ src/Binding/StoppableBinder.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace WPDesk\Init\Binding;
-
-interface StoppableBinder extends Hookable {
-
-	public function should_stop(): bool;
-}