+++ src/Module/BuiltinModule.php
@@ -4,6 +4,7 @@ declare( strict_types=1 );
 namespace WPDesk\Init\Module;
 
 use Psr\Container\ContainerInterface;
+use WPDesk\Init\Binding\Loader\EmptyDefinitions;
 use WPDesk\Init\Binding\Loader\BindingDefinitions;
 use WPDesk\Init\Binding\Loader\FilesystemDefinitions;
 use WPDesk\Init\Bootstrap\BootstrapContext;
@@ -17,6 +18,14 @@ final class BuiltinModule implements Module {
 
 	public function bindings( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions {
 		return new FilesystemDefinitions( __DIR__ . '/../Resources/bindings' );
+	}
+
+	public function activation( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions {
+		return new EmptyDefinitions();
+	}
+
+	public function deactivation( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions {
+		return new EmptyDefinitions();
 	}
 
 	public function gates( ContainerInterface $container, BootstrapContext $context ): array {