+++ src/Module/Module.php
@@ -15,6 +15,10 @@ interface Module {
 
 	public function bindings( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions;
 
+	public function activation( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions;
+
+	public function deactivation( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions;
+
 	/**
 	 * @return BootGate[]
 	 */