+++ src/Module/RequirementsModule.php
@@ -6,6 +6,7 @@ namespace WPDesk\Init\Module;
 use Psr\Container\ContainerInterface;
 use WPDesk\Init\Binding\Loader\ArrayDefinitions;
 use WPDesk\Init\Binding\Loader\BindingDefinitions;
+use WPDesk\Init\Binding\Loader\EmptyDefinitions;
 use WPDesk\Init\Bootstrap\BootstrapContext;
 use WPDesk\Init\Bootstrap\RequirementsGate;
 use WPDesk\Init\DependencyInjection\ContainerBuilder;
@@ -22,13 +23,17 @@ final class RequirementsModule implements Module {
 		return new ArrayDefinitions( [] );
 	}
 
+	public function activation( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions {
+		return new EmptyDefinitions();
+	}
+
+	public function deactivation( ContainerInterface $container, BootstrapContext $context ): BindingDefinitions {
+		return new EmptyDefinitions();
+	}
+
 	public function gates( ContainerInterface $container, BootstrapContext $context ): array {
-		$config = $context->module_config( self::class );
-		if ( isset( $config['requirements'] ) ) {
-			$requirements = $config['requirements'];
-		} else {
-			$requirements = $context->config()->get( 'requirements', [] );
-		}
+		$config       = $context->module_config( self::class );
+		$requirements = isset( $config['requirements'] ) ? $config['requirements'] : [];
 
 		return [ new RequirementsGate( $context->plugin(), (array) $requirements ) ];
 	}